CodeFix should ignore non-public properties as well, just like the analyzer.
